Some adjustments after new Python-version where float is not implicitly converted to int. Much more changes are probably required.

+++ tools/customplot.py
@@ -777,9 +777,9 @@ class plotsqlitewindow(QtWidgets.QMainWindow, customplot_ui_class):
             screen_dpi = QApplication.screens()[0].logicalDotsPerInch()
             width_pixels = width_inches * screen_dpi
             height_pixels = height_inches * screen_dpi
-            self.canvas.setFixedSize(width_pixels, height_pixels)
-            self.widgetPlot.setFixedWidth(max(self.canvas.size().width(), self.mpltoolbar.size().width()))
-            self.widgetPlot.setFixedHeight(self.canvas.size().height() + self.mpltoolbar.size().height()*3)
+            self.canvas.setFixedSize(int(width_pixels), int(height_pixels))
+            self.widgetPlot.setFixedWidth(int(max(self.canvas.size().width(), self.mpltoolbar.size().width())))
+            self.widgetPlot.setFixedHeight(int(self.canvas.size().height() + self.mpltoolbar.size().height()*3))
 
     def storesettings(self):
         self.ms.settingsdict['custplot_table1'] = str(self.table_ComboBox_1.currentText())
